调整main.py结构1.0
将调用数据库和调出当日文章的相关代码整合出来
